+++ contracts/core/CoreInternal.sol
@@ -16,8 +16,8 @@
 
 pragma solidity 0.4.24;
 
-import { CoreState } from "./lib/CoreState.sol";
 import { Ownable } from "zeppelin-solidity/contracts/ownership/Ownable.sol";
+import { CoreState } from "./lib/CoreState.sol";
 
 
 /**
